added like num to card
