+++ react-app/src/App.js
@@ -10,7 +10,7 @@ import { authenticate } from './store/session';
 import LoginPage from './components/splash/LoginPage';
 import SignUpPage from './components/splash/SignUpPage';
 import CreatePostForm from './components/Post/CreatePost/CreatePostForm';
-import GetUserPosts from './components/Post/GetUserPosts/GetUserPosts'
+import GetUserPosts from './components/UserProfile/GetUserPosts/GetUserPosts'
 
 
 import FeedPage from './components/Feed/FeedPage';
