+++ react-app/src/components/Post/PostDetails/index.js
@@ -18,8 +18,8 @@ function PostDetailsModal({post}) {
   return (
     <div className='profile-post-container'>
       <div className="post-icon-container">
-        <i className="fas fa-heart hover-icon"></i>
-        <i className="fas fa-comment hover-icon"></i>
+        <i className="fas fa-heart hover-icon"><span>{post?.comments?.length}</span></i>
+        <i className="fas fa-comment hover-icon"><span>{post?.likes?.length}</span></i>
       </div>
       <img className='profile-post-img' src={post?.photos[0]?.photo} onClick={() => setShowModal(true)}></img>
       {showModal && (
