+++ react-app/src/components/UserProfile/UserProfile.css
@@ -1,22 +1,54 @@
-.user-profile-container {
-  background-color: pink;
+.profile-heading-container {
   font-size: 2rem;
+  display: grid;
+  /* justify-content: center; */
+  /* grid-template-columns: minmax(30rem, 1fr) minmax(60rem 2fr); */
+  grid-template-columns: 1fr 2fr;
+  /* row-gap: 3rem; */
+  column-gap: 1.6rem;
+  margin-bottom: 3rem;
 }
+
 
 .profile-pic {
-  background-color: aqua;
-  border-radius: 10rem;
-  height: 15rem;
-  width: 15rem;
+  border-radius: 50%;
+  height: 20rem;
+  width: 20rem;
   margin: 2rem 5rem;
+  object-fit: cover;
+}
+.profile-info-container {
+
+  display: flex;
+  flex-direction: column;
+  /* width: 65%; */
 }
 
-.profile-info-container {
-  height: 15rem;
-  width: 15rem;
-  margin: 2rem 5rem;
-  background-color: brown;
+.profile-info-container > div {
+  height: 28%;
+
 }
+
+.username-profile-heading {
+  padding-top: 1.5rem;
+  font-size: 3rem;
+}
+
+.user-info-profile-heading,
+.full-name-profile-heading {
+  font-size: 1.8rem;
+}
+
+.full-name-profile-heading {
+  font-weight: bolder;
+}
+
+
+.profile-heading-container span:not(:first-child) {
+  padding-left: 3.5rem;
+}
+
+
 
 .profile-feed-container {
   display: grid;
@@ -61,9 +93,26 @@
   font-size: 2rem;
   margin: 0 2rem;
   cursor: pointer;
+
   /* position: absolute; */
 }
 
 .profile-post-container:hover .hover-icon {
   z-index: 1;
+}
+
+.profile-section-border {
+  width: 100%;
+  height: 5rem;
+  /* margin-top: 4rem; */
+  border-top: var(--border-light)
+
+}
+
+.all-container {
+  padding-top: 3rem;
+}
+
+.post-icon-container span {
+  padding-left: 1rem;
 }
