+++ react-app/src/index.css
@@ -20,6 +20,7 @@ body {
   background-color: rgb(250, 250, 250);
   display: flex;
   justify-content: center;
+  font-family: Arial, Helvetica, sans-serif;
 }
 
 #root {
